requirements.txt
requests
tqdm
urllib3
python-dotenv
pyinstaller
PyQt5